+++ rpc/triple/pb/dubbogo-grpc/dubbogo-server/cmd/server.go
@@ -19,6 +19,7 @@ package main
 
 import (
 	"fmt"
+	"github.com/apache/dubbo-go-samples/rpc/triple/pb/dubbogo-grpc/dubbogo-server/pkg"
 	"os"
 	"os/signal"
 	"syscall"
@@ -31,10 +32,6 @@ import (
 	_ "dubbo.apache.org/dubbo-go/v3/imports"
 
 	_ "github.com/dubbogo/triple/pkg/triple"
-)
-
-import (
-	"github.com/apache/dubbo-go-samples/rpc/triple/pb/dubbogo-grpc/server/dubbogo-server/pkg"
 )
 
 var (
