--- rpc/triple/pb/dubbogo-grpc/unary-client/dubbogo-client/conf/dubbogo.yml
@@ -1,10 +1,0 @@
-dubbo:
-  registries:
-    zk:
-      address: zookeeper://127.0.0.1:2181
-  consumer:
-    registry: zk
-    references:
-      greeterImpl:
-        protocol: tri
-        interface: org.apache.dubbo.Greeter
